deleted examples, fixed angle error in leg inv kin
the inverse kinematics of the leg used a wrong convention to calculate the relative position of chassis and leg. This was fixed.

Additionally the position control simulation was changed to the full model. 

Since the libraries where deleted the link to the spider walking was broken in the joystick walking sim